chore(builder): move constants and utils to shared package
