+++ packages/builder/builder-shared/src/deepFreezed.ts
@@ -1,0 +1,12 @@
+import assert from 'assert';
+import _ from '@modern-js/utils/lodash';
+
+export function deepFreezed<T extends Record<any, any> | any[]>(obj: T): T {
+  assert(typeof obj === 'object');
+  const handle = (item: any) =>
+    typeof item === 'object' ? deepFreezed(item) : item;
+  const ret = (
+    Array.isArray(obj) ? _.map(obj, handle) : _.mapValues(obj, handle)
+  ) as T;
+  return Object.freeze(ret);
+}
