+++ packages/builder/builder-shared/src/index.ts
@@ -1,7 +1,13 @@
-export * from './types';
-export * from './utils';
-export * from './logger';
+export * from './applyDefaultBuilderOptions';
+export * from './constants';
 export * from './createHook';
 export * from './createPluginStore';
-export * from './mergeBuilderConfig';
 export * from './createPublicContext';
+export * from './deepFreezed';
+export * from './fs';
+export * from './getBrowserslist';
+export * from './logger';
+export * from './mergeBuilderConfig';
+export * from './pick';
+export * from './regexp';
+export * from './types';
