+++ packages/builder/builder-shared/src/pick.ts
@@ -1,0 +1,8 @@
+export function pick<T, U extends keyof T>(obj: T, keys: ReadonlyArray<U>) {
+  return keys.reduce((ret, key) => {
+    if (obj[key] !== undefined) {
+      ret[key] = obj[key];
+    }
+    return ret;
+  }, {} as Pick<T, U>);
+}
