+++ packages/builder/builder-shared/src/types/utils.ts
@@ -10,6 +10,8 @@ export type ArrayOrNot<T> = T | T[];
 
 export type PromiseOrNot<T> = T | Promise<T>;
 
+export type NodeEnv = 'development' | 'production' | 'test';
+
 export type ChainedConfig<Config, Utils = unknown> = ArrayOrNot<
   | Config
   | (keyof Utils extends never
