+++ packages/builder/builder-webpack-provider/src/plugins/css.ts
@@ -1,5 +1,5 @@
-import { CSS_REGEX, getBrowserslistWithDefault } from '../shared';
-import type { BuilderContext } from '@modern-js/builder-shared';
+import { CSS_REGEX, type BuilderContext } from '@modern-js/builder-shared';
+import { getBrowserslistWithDefault } from '../shared';
 import {
   BuilderConfig,
   BuilderPlugin,
