+++ packages/builder/builder-webpack-provider/src/plugins/fallback.ts
@@ -1,4 +1,4 @@
-import { JS_REGEX, TS_REGEX } from '../shared';
+import { JS_REGEX, TS_REGEX } from '@modern-js/builder-shared';
 import type { BuilderPlugin } from '../types';
 import type { RuleSetRule } from 'webpack';
