+++ packages/builder/builder-webpack-provider/src/plugins/image.ts
@@ -1,11 +1,6 @@
 import { join } from 'path';
-import {
-  getDistPath,
-  getFilename,
-  getRegExpForExts,
-  IMAGE_EXTENSIONS,
-  getDataUrlCondition,
-} from '../shared';
+import { getRegExpForExts, IMAGE_EXTENSIONS } from '@modern-js/builder-shared';
+import { getDistPath, getFilename, getDataUrlCondition } from '../shared';
 import type { BuilderPlugin } from '../types';
 
 export const PluginImage = (): BuilderPlugin => ({
