+++ packages/builder/builder-webpack-provider/src/plugins/less.ts
@@ -1,5 +1,5 @@
-import { LESS_REGEX } from '../shared';
-import { BuilderPlugin, LessLoaderOptions } from '../types';
+import { LESS_REGEX } from '@modern-js/builder-shared';
+import type { BuilderPlugin, LessLoaderOptions } from '../types';
 
 export function PluginLess(): BuilderPlugin {
   return {
