+++ packages/builder/builder-webpack-provider/src/plugins/splitChunks.ts
@@ -1,12 +1,14 @@
+import assert from 'assert';
+import { RUNTIME_CHUNK_NAME } from '@modern-js/builder-shared';
+import { getPackageNameFromModulePath } from '../shared';
+
 import type { BuilderPlugin } from '../types';
 import type {
   BuilderChunkSplit,
   SplitChunks,
   CacheGroup,
 } from '../types/config/performance';
-import assert from 'assert';
 import type { Module } from 'webpack';
-import { getPackageNameFromModulePath, RUNTIME_CHUNK_NAME } from '../shared';
 
 // We expose the three-layer to specify webpack chunk-split ability:
 // 1. By strategy.There some best pratice integrated in our internal strategy.
