+++ packages/builder/builder-webpack-provider/src/plugins/tsLoader.ts
@@ -1,4 +1,4 @@
-import { TS_REGEX } from '../shared';
+import { TS_REGEX } from '@modern-js/builder-shared';
 import { BuilderPlugin } from '../types';
 import { applyScriptCondition, getUseBuiltIns } from './babel';
