+++ packages/builder/builder-webpack-provider/src/shared/index.ts
@@ -1,5 +1,3 @@
 export * from './fs';
 export * from './utils';
 export * from './format';
-export * from './constants';
-export * from './browserslist';
