+++ packages/builder/builder-webpack-provider/src/shared/utils.ts
@@ -1,52 +1,30 @@
-import assert from 'assert';
-import type * as webpack from 'webpack';
 import { URLSearchParams } from 'url';
+import {
+  getBrowserslist,
+  MODULE_PATH_REGEX,
+  DEFAULT_BROWSERSLIST,
+  DEFAULT_DATA_URL_SIZE,
+} from '@modern-js/builder-shared';
+
 import type Buffer from 'buffer';
-import { DEFAULT_DATA_URL_SIZE } from './constants';
+import type webpack from 'webpack';
 import type { SomeJSONSchema } from '@modern-js/utils/ajv/json-schema';
-import { BuilderConfig, DataUriLimit } from '../types';
+import type { BuilderConfig, DataUriLimit } from '../types';
 
-export const JS_REGEX = /\.(js|mjs|cjs|jsx)$/;
-export const TS_REGEX = /\.(ts|mts|cts|tsx)$/;
-export const CSS_REGEX = /\.css$/;
-export const LESS_REGEX = /\.less$/;
-export const SASS_REGEX = /\.s(a|c)ss$/;
-export const CSS_MODULE_REGEX = /\.module\.css$/;
-export const GLOBAL_CSS_REGEX = /\.global\.css$/;
-export const NODE_MODULES_REGEX = /node_modules/;
-export const SVG_REGEX = /\.svg$/;
-export const MODULE_PATH_REGEX =
-  /[\\/]node_modules[\\/](\.pnpm[\\/])?(?:(@[^[\\/]+)(?:[\\/]))?([^\\/]+)/;
+export async function getBrowserslistWithDefault(
+  path: string,
+  config: BuilderConfig,
+) {
+  if (config?.output?.overrideBrowserslist) {
+    return config.output.overrideBrowserslist;
+  }
 
-export const isNodeModulesCss = (path: string) =>
-  NODE_MODULES_REGEX.test(path) &&
-  CSS_REGEX.test(path) &&
-  !CSS_MODULE_REGEX.test(path);
-
-export const mergeRegex = (...regexes: (string | RegExp)[]): RegExp => {
-  assert(
-    regexes.length,
-    'mergeRegex: regular expression array should not be empty.',
-  );
-  const sources = regexes.map(regex =>
-    regex instanceof RegExp ? regex.source : regex,
-  );
-
-  return new RegExp(sources.join('|'));
-};
+  const result = await getBrowserslist(path);
+  return result || DEFAULT_BROWSERSLIST;
+}
 
 /** Preserving the details of schema by generic types. */
 export const defineSchema = <T extends SomeJSONSchema>(schema: T): T => schema;
-
-export function getRegExpForExts(extensions: string[]): RegExp {
-  return new RegExp(
-    `\\.(${extensions
-      .map(ext => ext.trim())
-      .map(ext => (ext.startsWith('.') ? ext.slice(1) : ext))
-      .join('|')})$`,
-    'i',
-  );
-}
 
 export const getDataUrlLimit = (
   config: BuilderConfig,
